+++ Sources/MapboxMaps/Ornaments/DistanceFormatter.swift
@@ -11,20 +11,22 @@ internal class DistanceFormatter: MeasurementFormatter {
     ///
     /// - parameter distance: The distance, measured in meters.
     /// - returns: A localized formatted distance string including units.
-    internal func string(fromDistance distance: CLLocationDistance) -> String {
+    internal func string(fromDistance distance: CLLocationDistance, useMetricSystem: Bool? = nil) -> String {
 
         numberFormatter.roundingIncrement = 0.25
 
         var measurement = Measurement(value: distance, unit: UnitLength.meters)
 
-        if !locale.usesMetricSystem {
+        let shouldUseMetricSystem = useMetricSystem ?? locale.usesMetricSystem
+
+        if shouldUseMetricSystem {
+            unitOptions = [.providedUnit, .naturalScale]
+        } else {
             unitOptions = .providedUnit
             measurement.convert(to: .miles)
             if measurement.value <= 0.2 {
                 measurement.convert(to: .feet)
             }
-        } else {
-            unitOptions = .naturalScale
         }
         return string(from: measurement)
     }